Updated config files and added mdx plugins, switched to pnpm
