--- next.config.js
@@ -1,19 +1,0 @@
-const withMDX = require('@next/mdx')();
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: 'https',
-        hostname: 'ik.imagekit.io',
-      },
-    ],
-  },
-  pageExtensions: ['js', 'jsx', 'mdx', 'ts', 'tsx'],
-  experimental: {
-    mdxRs: true,
-  },
-};
-
-module.exports = withMDX(nextConfig);
