+++ package.json
@@ -21,7 +21,6 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@react-email/components": "^0.0.22",
     "@react-email/tailwind": "^0.0.19",
-    "autoprefixer": "10.4.19",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "framer-motion": "^11.3.21",
@@ -43,17 +42,23 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.14",
     "@types/mdx": "^2.0.13",
     "@types/node": "^22.0.2",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@vercel/style-guide": "^6.0.0",
+    "autoprefixer": "^10.4.20",
     "eslint": "8.56.0",
     "eslint-config-next": "14.2.5",
     "postcss": "8.4.40",
     "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.5",
     "prisma": "^5.17.0",
+    "rehype-autolink-headings": "^7.1.0",
+    "rehype-pretty-code": "^0.13.2",
+    "rehype-slug": "^6.0.0",
+    "remark-gfm": "^4.0.0",
     "tailwindcss": "3.4.7",
     "typescript": "5.5.4"
   }
