+++ src/content/projects/grapple-grub.mdx
@@ -1,26 +1,15 @@
 ---
-title: 'Icon.io'
-date: '2023-10-28'
+title: 'Grapple Grub'
+date: '2023-3-20'
 featured: true
-role: 'Full-Stack Developer'
-summary: 'Icon.io is a dynamic online multiplayer platform aiming to deliver a fun and interactive gaming experience for you and your friends!'
-thumbnailUrl: 'https://ik.imagekit.io/ryangan/Project%20Screenshots/iconio/icon.io-thumbnail.png?updatedAt=1723238598008'
-stack:
-  [
-    'react',
-    'nodejs',
-    'typescript',
-    'socketio',
-    'figma',
-    'antdesign',
-    'css3',
-    'aws',
-    'pm2',
-  ]
+role: 'Gameplay Developer'
+summary: 'Swinging and slinging grub, all across the city!'
+thumbnailUrl: 'https://ik.imagekit.io/ryangan/Project%20Screenshots/grapple-grub/grapplegrub-thumbnail.png?updatedAt=1723239351853'
+stack: ['unity', 'csharp', 'adobeps', 'adobeai', 'trello']
 link:
   {
     github: 'https://github.com/ryangandev/icon.io',
-    live: 'https://icon.ryiscrispy.com/',
+    live: 'https://sites.google.com/view/hpfishingco',
   }
 ---
 
@@ -41,3 +30,7 @@ Plans are underway to introduce more games soon, such as the classic Tic-Tac-Toe
 - Implemented the **whiteboard** component using the **Canvas API** and **Socket.IO** to ensure synchronized broadcasting of drawings in real-time
 
 - Configured and deployed the application on **AWS EC2** and utilized the process manager **PM2** to ensure continuous uptime and availability
+
+```tsx
+const spinner = new Spinner();
+```
