+++ src/content/projects/player2-helpdesk.mdx
@@ -1,10 +1,10 @@
 ---
 title: 'Player 2 Helpdesk'
-date: '2022-10-13'
+date: '2023-5-24'
 featured: true
 role: 'Frontend Developer'
 summary: 'Developed for Dynasty11 Studio, the "Player 2 Helpdesk" application serves as the primary support channel for users of the "Player 2" social networking platform'
-thumbnailUrl: 'https://ik.imagekit.io/ryangan/player2-thumbnail.png?updatedAt=1695996409769'
+thumbnailUrl: 'https://ik.imagekit.io/ryangan/Project%20Screenshots/player2-helpdesk/player2-thumbnail.png?updatedAt=1723238543464'
 stack:
   [
     'react',
@@ -21,7 +21,7 @@ stack:
 link: { live: 'https://player2app.com/' }
 ---
 
-### Overview
+## Overview
 
 Player 2 Helpdesk is a custom web application tailored to support the "Player 2" social networking community. It facilitates efficient communication between the Player 2 app users and the support staffs.
 
@@ -29,12 +29,12 @@ By bridging the gap between users and support staffs, it ensures efficient commu
 
 ## Responsibilities
 
-- Led the frontend team from conceptualizing the UI in **Figma** to translating the design into a fully functional implementation using **React** and **TypeScript**
+1. Led the frontend team from conceptualizing the UI in **Figma** to translating the design into a fully functional implementation using **React** and **TypeScript**
 
-- Designed a user-friendly **dashboard** component with advanced filtering and sorting capabilities, ensuring efficient ticket management for the support team
+2. Designed a user-friendly **dashboard** component with advanced filtering and sorting capabilities, ensuring efficient ticket management for the support team
 
-- Developed a dynamic **reporting system** that supports diverse date range and filter selections, utilizing **Rechart**, to offer insights into helpdesk performance
+3. Developed a dynamic **reporting system** that supports diverse date range and filter selections, utilizing **Rechart**, to offer insights into helpdesk performance
 
-- Collaborated closely with the backend team to harness **server-sent events**, enabling real-time chat and notification features
+4. Collaborated closely with the backend team to harness **server-sent events**, enabling real-time chat and notification features
 
-- Configured and managed application state using **Redux** to ensure state consistency across various functionalities
+5. Configured and managed application state using **Redux** to ensure state consistency across various functionalities
